+++ CHANGES.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+### 2.?.? - 2020-??-??
+
+* Fixed writing duplicate resource that are referenced multiple times. [#483](https://github.com/AnalyticalGraphicsInc/gltf-pipeline/pull/483)
+
 ### 2.1.8 - 2020-03-04
 
 * Fixed a bug in `processGltf` and `gltfToGlb` where the user's `options` object was getting modified with internal options. [#528](https://github.com/CesiumGS/gltf-pipeline/pull/528)